+++ response/file.go
@@ -1,0 +1,37 @@
+package response
+
+import (
+	"io/ioutil"
+	"net/http"
+	"os"
+	"strings"
+)
+
+func File(filepath string) *Response {
+	var err error
+	f, err := os.Open(filepath)
+	if err != nil {
+		return New("").Status(http.StatusNotFound)
+	}
+	b, err := ioutil.ReadAll(f)
+	if err != nil {
+		return New("").Status(http.StatusUnprocessableEntity)
+	}
+	resp := New(string(b))
+	fileSuffix := filepath[strings.LastIndexByte(filepath, '.')+1:]
+	resp.Headers(Headers{
+		"Content-Type": defaultSuffixMapToContentTypes[fileSuffix],
+	})
+	return resp
+}
+
+var defaultSuffixMapToContentTypes = map[string]string{
+	"html": "text/html",
+	"css":  "text/css",
+	"js":   "application/javascript",
+	"json": "application/json",
+	"xml":  "application/xml",
+	"gif":  "image/gif",
+	"png":  "image/png",
+	"jpeg": "image/jpeg",
+}