+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 - fix: wildcard route matching
 - `func (*cookie.Cookie) MaxAge(int)`: use to modified max age field of the cookie
 - add request fairing
+- feat: `func File(filepath string) *Response` at package `response`, use to create a file response with default content-type
 
 ## v0.12.9